requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
ansible==7.4.0
ansible-core==2.14.4
async-timeout==4.0.2
attrs==23.1.0
Brotli==1.0.9
cachelib==0.10.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.0.7
cryptography==40.0.1
cycler==0.11.0
databricks-cli==0.17.6
dataclasses-json==0.5.7
docker==6.0.1
entrypoints==0.4
Flask==2.2.3
Flask-Session==0.5.0
fonttools==4.39.2
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
langchain==0.0.200
langchainplus-sdk==0.0.10
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
mlflow==2.2.2
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.7
numba==0.56.4
numexpr==2.8.4
numpy==1.23.0
oauthlib==3.2.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.0
pandas==2.0.2
pdfminer==20191125
Pillow==9.5.0
pip==23.0.1
protobuf==4.23.1
pyarrow==11.0.0
pycparser==2.21
pycryptodome==3.18.0
pydantic==1.10.8
PyJWT==2.6.0
pyodbc==4.0.39
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
PyYAML==6.0
querystring-parser==1.2.4
rank-bm25==0.2.2
regex==2022.7.9
requests==2.31.0
resolvelib==0.8.1
scikit-learn==1.2.2
scipy==1.10.1
setuptools==66.0.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
SQLAlchemy==2.0.7
sqlparse==0.4.3
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tqdm==4.65.0
typing_extensions==4.5.0
typing-inspect==0.9.0
tzdata==2023.3
urllib3==2.0.2
waitress==2.1.2
watchdog==3.0.0
websocket-client==1.5.1
Werkzeug==2.2.3
wheel==0.38.4
yarl==1.9.2
zipp==3.15.0